content-safety/python/requirements.txt
azure-ai-contentsafety==1.0.0
azure-core==1.29.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
